# GH-3933 Thread safe MemStatementList