--- a/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/MemLiteral.java
+++ b/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/MemLiteral.java
@@ -100,34 +100,27 @@ public class MemLiteral extends SimpleLiteral implements MemValue {
 
 	@Override
 	public MemStatementList getObjectStatementList() {
-
 		return objectStatements;
-
 	}
 
 	@Override
 	public int getObjectStatementCount() {
-
 		return objectStatements.size();
-
 	}
 
 	@Override
 	public void addObjectStatement(MemStatement st) throws InterruptedException {
-
 		objectStatements.add(st);
 	}
 
 	@Override
 	public void removeObjectStatement(MemStatement st) throws InterruptedException {
 		objectStatements.remove(st);
-
 	}
 
 	@Override
 	public void cleanSnapshotsFromObjectStatements(int currentSnapshot) throws InterruptedException {
 		objectStatements.cleanSnapshots(currentSnapshot);
-
 	}
 
 	@Override