--- a/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/MemStatementList.java
+++ b/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/MemStatementList.java
@@ -7,161 +7,254 @@
  *******************************************************************************/
 package org.eclipse.rdf4j.sail.memory.model;
 
-import java.util.Arrays;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.VarHandle;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.concurrent.locks.LockSupport;
+import java.util.concurrent.locks.StampedLock;
 
 /**
  * A dedicated data structure for storing MemStatement objects, offering operations optimized for their use in the
  * memory Sail.
  */
 public class MemStatementList {
-	private static final MemStatement[] EMPTY_ARRAY = new MemStatement[0];
+	private static final MemStatement[] EMPTY_ARRAY = {};
 
-	/*-----------*
-	 * Variables *
-	 *-----------*/
-
-	private volatile MemStatement[] statements;
+	// statements will be null when the array is growing
+	private volatile MemStatement[] statements = EMPTY_ARRAY;
+	private static final VarHandle STATEMENTS;
+	static final VarHandle STATEMENTS_ARRAY;
 
 	private volatile int size;
+	private static final VarHandle SIZE;
 
-	/*--------------*
-	 * Constructors *
-	 *--------------*/
+	// When inserting a new statement into the statements array we need to iterate through the array looking for a free
+	// spot. We keep track of where we last inserted a statement by storting that index in previouslyInsertedIndex. This
+	// doesn't guarantee that previouslyInsertedIndex+1 wil be free, but it gives us a decent hint as to where to start
+	// looking. When multiple threads are inserting at the same time the previouslyInsertedIndex should be considered
+	// "best effort".
+	private volatile int previouslyInsertedIndex = -1;
+	private static final VarHandle PREVIOUSLY_INSERTED_INDEX;
 
-	/**
-	 * Creates a new MemStatementList.
-	 */
+	private volatile int guaranteedLastIndexInUse = -1;
+	private static final VarHandle GUARANTEED_LAST_INDEX_IN_USE;
+
+	private volatile boolean prioritiseCleanup;
+	private static final VarHandle PRIORITISE_CLEANUP;
+
+	private final StampedLock addRemoveLock = new StampedLock();
+
+	private final AtomicReference<Thread> prioritisedThread = new AtomicReference<>();
+
 	public MemStatementList() {
-		statements = EMPTY_ARRAY;
-		size = 0;
 	}
 
 	public MemStatementList(int capacity) {
 		statements = new MemStatement[capacity];
-		size = 0;
 	}
-
-	public MemStatementList(MemStatementList other) {
-		statements = Arrays.copyOf(other.statements, other.statements.length);
-		size = other.size;
-	}
-
-	/*---------*
-	 * Methods *
-	 *---------*/
 
 	public int size() {
-		return size;
+		return ((int) SIZE.getAcquire(this));
 	}
 
 	public boolean isEmpty() {
-		return size == 0;
+		return ((int) SIZE.getAcquire(this)) == 0;
 	}
 
-	public MemStatement get(int index) {
-		assert index >= 0 : "index < 0";
-		assert index < size : "index >= size";
+	public void add(MemStatement st) throws InterruptedException {
 
-		return statements[index];
-	}
-
-	public MemStatement getIfExists(int index) {
-		if (index < size) {
-			return statements[index];
-		} else {
-			return null;
-		}
-	}
-
-	public void add(MemStatement st) {
-		if (size == statements.length) {
-			// Grow array
-			growArray((size == 0) ? 4 : 2 * size);
-		}
-
-		statements[size] = st;
-		++size;
-	}
-
-	public void addAll(MemStatementList other) {
-		if (size + other.size >= statements.length) {
-			// Grow array
-			growArray(size + other.size);
-		}
-
-		System.arraycopy(other.statements, 0, statements, size, other.size);
-		size += other.size;
-	}
-
-	public void remove(int index) {
-		assert index >= 0 : "index < 0";
-		assert index < size : "index >= size";
-
-		if (index == size - 1) {
-			// Last statement in array
-			statements[index] = null;
-			--size;
-		} else {
-			// Not last statement in array, move last
-			// statement over the one at [index]
-			--size;
-			statements[index] = statements[size];
-			statements[size] = null;
-		}
-	}
-
-	public void remove(MemStatement st) {
-		for (int i = 0; i < size; ++i) {
-			if (statements[i] == st) {
-				remove(i);
-				return;
+		if (((boolean) PRIORITISE_CLEANUP.getOpaque(this))) {
+			long start = System.currentTimeMillis();
+			long stop = start + TimeUnit.SECONDS.toMillis(30);
+			while (stop > System.currentTimeMillis() && ((boolean) PRIORITISE_CLEANUP.getVolatile(this))) {
+				LockSupport.parkNanos(TimeUnit.MILLISECONDS.toNanos(10));
 			}
 		}
+
+		long readLock = addRemoveLock.readLock();
+		try {
+			do {
+
+				MemStatement[] statements = getStatements();
+				int length = statements.length;
+
+				if (length > (int) SIZE.getAcquire(this)) {
+
+					int previouslyInsertedIndex = (int) PREVIOUSLY_INSERTED_INDEX.getOpaque(this);
+					if (previouslyInsertedIndex > length) {
+						continue;
+					}
+
+					int i = previouslyInsertedIndex + 1 == length ? 0 : previouslyInsertedIndex + 1;
+
+					for (; i != previouslyInsertedIndex; i = i + 1 == length ? 0 : i + 1) {
+
+						if (statements[i] == null) {
+
+							boolean success = STATEMENTS_ARRAY.compareAndSet(statements, i, null, st);
+							if (success) {
+
+								// check if the statements array has been swapped out (because it was grown) while we
+								// were
+								// inserting into it
+								MemStatement[] statementsAfterInsert = getStatements();
+								if (statementsAfterInsert != statements
+										&& STATEMENTS_ARRAY.getAcquire(statements, i) != st) {
+									// we wrote into an array while it was growing and our write was lost
+									break;
+								}
+
+								PREVIOUSLY_INSERTED_INDEX.setRelease(this, i);
+								SIZE.getAndAdd(this, 1);
+
+								updateGuaranteedLastIndexInUse(i);
+
+								return;
+							}
+						}
+
+					}
+
+				}
+
+				// statements array is probably full
+
+				if (STATEMENTS.compareAndSet(this, statements, null)) {// Grow array
+					MemStatement[] newArray = new MemStatement[Math.max(4, length * 2)];
+					if (statements != EMPTY_ARRAY) {
+						System.arraycopy(statements, 0, newArray, 0, length);
+					}
+
+					STATEMENTS.setRelease(this, newArray);
+				}
+				if (Thread.interrupted()) {
+					throw new InterruptedException();
+				}
+			} while (true);
+		} finally {
+			addRemoveLock.unlockRead(readLock);
+		}
+	}
+
+	private void updateGuaranteedLastIndexInUse(int newValue) {
+		int guaranteedLastIndexInUse = (int) GUARANTEED_LAST_INDEX_IN_USE.getAcquire(this);
+		if (guaranteedLastIndexInUse < newValue) {
+			while (guaranteedLastIndexInUse < newValue
+					&& !GUARANTEED_LAST_INDEX_IN_USE.compareAndSet(this, guaranteedLastIndexInUse, newValue)) {
+				guaranteedLastIndexInUse = (int) GUARANTEED_LAST_INDEX_IN_USE.getAcquire(this);
+			}
+
+		}
+	}
+
+	public void remove(MemStatement st) throws InterruptedException {
+
+		do {
+			MemStatement[] statements = getStatements();
+
+			boolean success = true;
+			for (int i = 0; i < statements.length; i++) {
+				if (statements[i] == st) {
+
+					success = innerRemove(st, statements, i);
+
+					break;
+				}
+			}
+			if (success) {
+				break;
+			}
+			if (Thread.interrupted()) {
+				throw new InterruptedException();
+			}
+		} while (true);
+	}
+
+	public void remove(MemStatement st, int index) throws InterruptedException {
+
+		do {
+			MemStatement[] statements = getStatements();
+
+			if (statements[index] == st && innerRemove(st, statements, index)) {
+				return;
+			} else {
+				remove(st);
+			}
+			if (Thread.interrupted()) {
+				throw new InterruptedException();
+			}
+		} while (true);
+	}
+
+	private boolean innerRemove(MemStatement st, MemStatement[] statements, int i) throws InterruptedException {
+		long writeLock = addRemoveLock.writeLock();
+		try {
+			if (getStatements() != statements) {
+				return false;
+			}
+
+			boolean success = STATEMENTS_ARRAY.compareAndSet(statements, i, st, null);
+			if (success) {
+				while (true) {
+					int size = size();
+					boolean decrementedSize = SIZE.compareAndSet(this, size, size - 1);
+					if (decrementedSize) {
+						return true;
+					}
+				}
+			} else {
+				return false;
+			}
+		} finally {
+			addRemoveLock.unlockWrite(writeLock);
+		}
+
 	}
 
 	public void clear() {
-		statements = EMPTY_ARRAY;
-		size = 0;
-	}
-
-	public void cleanSnapshots(int currentSnapshot) {
-		int i = size - 1;
-
-		// remove all deprecated statements from the end of the list
-		for (; i >= 0; i--) {
-			if (Thread.currentThread().isInterrupted()) {
-				return;
-			}
-			if (statements[i].getTillSnapshot() <= currentSnapshot) {
-				--size;
-				statements[i] = null;
-			} else {
-				i--;
-				break;
-			}
-		}
-
-		// remove all deprecated statements that are not at the end of the list
-		for (; i >= 0; i--) {
-			if (Thread.currentThread().isInterrupted()) {
-				return;
-			}
-
-			if (statements[i].getTillSnapshot() <= currentSnapshot) {
-				// replace statement with last statement in the list
-				--size;
-				statements[i] = statements[size];
-				statements[size] = null;
-			}
+		long writeLock = addRemoveLock.writeLock();
+		try {
+			statements = EMPTY_ARRAY;
+			size = 0;
+			previouslyInsertedIndex = -1;
+			guaranteedLastIndexInUse = -10;
+			prioritiseCleanup = false;
+		} finally {
+			addRemoveLock.unlockWrite(writeLock);
 		}
 	}
 
-	private void growArray(int newSize) {
-		MemStatement[] newArray = new MemStatement[newSize];
-		if (statements != EMPTY_ARRAY) {
-			System.arraycopy(statements, 0, newArray, 0, size);
-		}
-		statements = newArray;
+	public void cleanSnapshots(int currentSnapshot) throws InterruptedException {
+		boolean error;
+		do {
+			MemStatement[] statements = getStatements();
+
+			// reset the error flag
+			error = false;
+
+			for (int i = 0; i < statements.length; i++) {
+
+				if (Thread.interrupted()) {
+					throw new InterruptedException();
+				}
+
+				MemStatement statement = statements[i];
+				if (statement != null && statement.getTillSnapshot() <= currentSnapshot) {
+					boolean success = innerRemove(statement, statements, i);
+					if (!success) {
+						error = true;
+						break;
+					}
+				}
+
+			}
+
+			if (!error) {
+				// make sure that the statements list didn't grow while we were cleaning it
+				error = !STATEMENTS.compareAndSet(this, statements, statements);
+			}
+		} while (error);
 
 	}
 
@@ -178,20 +271,15 @@ public class MemStatementList {
 	 * @return
 	 */
 	public MemStatement getExact(MemResource subject, MemIRI predicate, MemValue object, MemResource context,
-			int snapshot) {
+			int snapshot) throws InterruptedException {
 
-		// these variables are volatile, so we store them locally to avoid having to do a lot of volatile reads
-		MemStatement[] statements = this.statements;
-		int size = this.size;
+		MemStatement[] statements = getStatements();
+		int lastIndexToCheck = getGuaranteedLastIndexInUse();
 
-		for (int i = 0; i < size; i++) {
+		for (int i = 0; i <= lastIndexToCheck; i++) {
 			MemStatement memStatement = statements[i];
-			if (memStatement == null) {
-				break;
-			}
-
-			// match predicate first, because the invoking method usually ends up with the subject list
-			if (memStatement.exactMatch(subject, predicate, object, context) && memStatement.isInSnapshot(snapshot)) {
+			if (memStatement != null && memStatement.exactMatch(subject, predicate, object, context)
+					&& memStatement.isInSnapshot(snapshot)) {
 				return memStatement;
 			}
 		}
@@ -204,7 +292,91 @@ public class MemStatementList {
 	 *
 	 * @return the underlying array og MemStatements
 	 */
-	MemStatement[] getStatements() {
+	public MemStatement[] getStatements() throws InterruptedException {
+		MemStatement[] statements = (MemStatement[]) STATEMENTS.getAcquire(this);
+		while (statements == null) {
+			if (Thread.interrupted()) {
+				throw new InterruptedException();
+			}
+			Thread.onSpinWait();
+			statements = (MemStatement[]) STATEMENTS.getAcquire(this);
+		}
 		return statements;
 	}
+
+	public int getGuaranteedLastIndexInUse() {
+		return ((int) GUARANTEED_LAST_INDEX_IN_USE.getAcquire(this));
+	}
+
+	public void setPrioritiseCleanup(boolean prioritiseCleanup) {
+		if (!prioritiseCleanup) {
+			if (prioritisedThread.compareAndSet(Thread.currentThread(), null)) {
+				PRIORITISE_CLEANUP.setVolatile(this, false);
+			} else {
+				assert !((boolean) PRIORITISE_CLEANUP.getVolatile(this));
+			}
+		} else {
+			if (prioritisedThread.compareAndSet(null, Thread.currentThread())) {
+				Thread.currentThread().setPriority(Thread.MAX_PRIORITY);
+				PRIORITISE_CLEANUP.setVolatile(this, true);
+			} else {
+				throw new IllegalStateException("A cleanup thread is already prioritised: " + prioritisedThread.get());
+			}
+		}
+	}
+
+	static {
+		try {
+			SIZE = MethodHandles.lookup()
+					.in(MemStatementList.class)
+					.findVarHandle(MemStatementList.class, "size", int.class);
+		} catch (ReflectiveOperationException e) {
+			throw new Error(e);
+		}
+	}
+
+	static {
+		try {
+			PREVIOUSLY_INSERTED_INDEX = MethodHandles.lookup()
+					.in(MemStatementList.class)
+					.findVarHandle(MemStatementList.class, "previouslyInsertedIndex", int.class);
+		} catch (ReflectiveOperationException e) {
+			throw new Error(e);
+		}
+	}
+
+	static {
+		try {
+			GUARANTEED_LAST_INDEX_IN_USE = MethodHandles.lookup()
+					.in(MemStatementList.class)
+					.findVarHandle(MemStatementList.class, "guaranteedLastIndexInUse", int.class);
+		} catch (ReflectiveOperationException e) {
+			throw new Error(e);
+		}
+	}
+
+	static {
+		try {
+			PRIORITISE_CLEANUP = MethodHandles.lookup()
+					.in(MemStatementList.class)
+					.findVarHandle(MemStatementList.class, "prioritiseCleanup", boolean.class);
+		} catch (ReflectiveOperationException e) {
+			throw new Error(e);
+		}
+	}
+
+	static {
+		try {
+			STATEMENTS = MethodHandles.lookup()
+					.in(MemStatementList.class)
+					.findVarHandle(MemStatementList.class, "statements", MemStatement[].class);
+		} catch (ReflectiveOperationException e) {
+			throw new Error(e);
+		}
+	}
+
+	static {
+		STATEMENTS_ARRAY = MethodHandles.arrayElementVarHandle(MemStatement[].class);
+	}
+
 }