GitHub issue resolved: #3933 <!-- add a Github issue number here, e.g #123. -->

Briefly describe the changes proposed in this PR:
 - Made MemStatementList thread-safe and removed read/write locks.
 - MemStatementList supports the following operations:
    - read concurrently with read/add/remove
    - add concurrently with read/add
    - remove and add are mutually exclusive, and also not concurrent with other remove operations (even though this should be technically possible)
 - Added functionality to track which snapshots can be cleaned.
 - Reduced the use of volatile reads/writes in MemStatement 

<!-- short description of your change goes here -->

----
PR Author Checklist (see the [contributor guidelines](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md) for more details):

 - [ ] my pull request is [self-contained](https://rdf4j.org/documentation/developer/merge-strategy/#self-contained-changes-pull-requests-and-commits)
 - [ ] I've added tests for the changes I made
 - [ ] I've applied [code formatting](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md#code-formatting) (you can use `mvn process-resources` to format from the command line)
 - [ ] I've [squashed](https://rdf4j.org/documentation/developer/squashing) my commits where necessary 
 - [ ] every commit message starts with the issue number (GH-xxxx) followed by a meaningful description of the change