@hmottestad 2022-05-25
Current status:
- [ ] Adding
- [ ] Removing
- [ ] Growing array
- [ ] Compacting array
- [ ] Shrinking array