@hmottestad 2022-05-25
### After
```
Benchmark                                                     Mode  Cnt      Score     Error  Units
QueryBenchmark.complexQuery                                   avgt    5      2.463 ±   0.125  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5      1.138 ±   0.071  ms/op
QueryBenchmark.groupByQuery                                   avgt    5      0.994 ±   0.046  ms/op
QueryBenchmark.long_chain                                     avgt    5    395.817 ±  11.158  ms/op
QueryBenchmark.lots_of_optional                               avgt    5     94.724 ±   1.024  ms/op
QueryBenchmark.minus                                          avgt    5  14990.039 ± 785.455  ms/op
QueryBenchmark.nested_optionals                               avgt    5    107.100 ±   0.212  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5      9.526 ±   0.702  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5      1.194 ±   0.035  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5     79.522 ±   9.886  ms/op
QueryBenchmark.simple_filter_not                              avgt    5      4.404 ±   0.759  ms/op
```