@hmottestad 2022-06-14
## Before
```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt    5  39.294 ± 1.146  ms/op
```

```
Benchmark                                                     Mode  Cnt     Score    Error  Units
QueryBenchmark.complexQuery                                   avgt    5     2.802 ±  0.023  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5     1.251 ±  0.007  ms/op
QueryBenchmark.groupByQuery                                   avgt    5     0.970 ±  0.004  ms/op
QueryBenchmark.long_chain                                     avgt    5   457.275 ±  3.869  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    97.808 ±  2.328  ms/op
QueryBenchmark.minus                                          avgt    5  2465.679 ± 49.315  ms/op
QueryBenchmark.nested_optionals                               avgt    5   106.406 ±  0.214  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     9.311 ±  0.062  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5     1.230 ±  0.020  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    75.272 ±  0.825  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     4.144 ±  0.008  ms/op
```

```
Benchmark                                                                           (isolationLevel)  Mode  Cnt     Score   Error  Units
LoadingBenchmark.loadRealData                                                                   NONE  avgt        370.741          ms/op
LoadingBenchmark.loadRealData                                                               SNAPSHOT  avgt        505.621          ms/op
LoadingBenchmark.loadRealData                                                           SERIALIZABLE  avgt        525.394          ms/op
LoadingBenchmark.loadSynthetic                                                                  NONE  avgt          6.124          ms/op
LoadingBenchmark.loadSynthetic                                                              SNAPSHOT  avgt          8.243          ms/op
LoadingBenchmark.loadSynthetic                                                          SERIALIZABLE  avgt          8.256          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                           NONE  avgt          6.798          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                       SNAPSHOT  avgt          8.741          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                   SERIALIZABLE  avgt          8.851          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                        NONE  avgt         12.929          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                    SNAPSHOT  avgt         34.652          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                SERIALIZABLE  avgt         37.664          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                           NONE  avgt       1267.861          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                       SNAPSHOT  avgt       1679.037          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                   SERIALIZABLE  avgt       1843.264          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                                NONE  avgt          6.213          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                            SNAPSHOT  avgt          8.648          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                        SERIALIZABLE  avgt          8.731          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                                    NONE  avgt         17.479          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                                SNAPSHOT  avgt         19.658          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                            SERIALIZABLE  avgt         20.308          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                                    NONE  avgt         18.628          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                                SNAPSHOT  avgt         27.683          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                            SERIALIZABLE  avgt         27.480          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                            NONE  avgt         12.416          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                        SNAPSHOT  avgt         16.548          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                    SERIALIZABLE  avgt         16.126          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing              NONE  avgt         11.756          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing          SNAPSHOT  avgt         28.789          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing      SERIALIZABLE  avgt         28.993          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                               NONE  avgt         17.590          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                           SNAPSHOT  avgt         19.517          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                       SERIALIZABLE  avgt         19.116          ms/op
```


## After
```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt    5  29.290 ± 1.547  ms/op
```

```
Benchmark                                                     Mode  Cnt     Score    Error  Units
QueryBenchmark.complexQuery                                   avgt    5     1.766 ±  0.015  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5     0.843 ±  0.011  ms/op
QueryBenchmark.groupByQuery                                   avgt    5     0.997 ±  0.015  ms/op
QueryBenchmark.long_chain                                     avgt    5   295.906 ±  5.224  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    62.895 ±  0.224  ms/op
QueryBenchmark.minus                                          avgt    5  2406.726 ± 51.509  ms/op
QueryBenchmark.nested_optionals                               avgt    5    84.027 ±  0.706  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     7.117 ±  0.067  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5     1.175 ±  0.009  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    75.663 ±  0.572  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     3.409 ±  0.005  ms/op
```

```
Benchmark                                                                           (isolationLevel)  Mode  Cnt     Score   Error  Units
LoadingBenchmark.loadRealData                                                                   NONE  avgt        392.139          ms/op
LoadingBenchmark.loadRealData                                                               SNAPSHOT  avgt        508.710          ms/op
LoadingBenchmark.loadRealData                                                           SERIALIZABLE  avgt        506.108          ms/op
LoadingBenchmark.loadSynthetic                                                                  NONE  avgt          7.572          ms/op
LoadingBenchmark.loadSynthetic                                                              SNAPSHOT  avgt          9.834          ms/op
LoadingBenchmark.loadSynthetic                                                          SERIALIZABLE  avgt          9.371          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                           NONE  avgt          7.955          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                       SNAPSHOT  avgt         10.098          ms/op
LoadingBenchmark.loadSyntheticAndSize                                                   SERIALIZABLE  avgt          9.760          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                        NONE  avgt         11.202          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                    SNAPSHOT  avgt         30.287          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransaction                                SERIALIZABLE  avgt         36.107          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                           NONE  avgt       1168.993          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                       SNAPSHOT  avgt       1438.994          ms/op
LoadingBenchmark.loadSyntheticOneStatementPerTransactionClearPrevious                   SERIALIZABLE  avgt       1596.147          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                                NONE  avgt          8.034          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                            SNAPSHOT  avgt         10.142          ms/op
LoadingBenchmark.loadSyntheticSingleTransactionGetFirstStatement                        SERIALIZABLE  avgt         10.895          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                                    NONE  avgt         16.556          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                                SNAPSHOT  avgt         19.716          ms/op
LoadingBenchmark.loadSyntheticWithDuplicates                                            SERIALIZABLE  avgt         19.795          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                                    NONE  avgt         17.095          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                                SNAPSHOT  avgt         26.960          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatements                            SERIALIZABLE  avgt         26.647          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                            NONE  avgt         12.956          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                        SNAPSHOT  avgt         16.900          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsGetFirst                    SERIALIZABLE  avgt         17.243          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing              NONE  avgt         13.109          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing          SNAPSHOT  avgt         30.342          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesAndNewStatementsIteratorMatchesNothing      SERIALIZABLE  avgt         29.353          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                               NONE  avgt         16.365          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                           SNAPSHOT  avgt         20.611          ms/op
LoadingBenchmark.loadSyntheticWithDuplicatesFlush                                       SERIALIZABLE  avgt         19.062          ms/op
```